backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
structlog>=24.2.0
pyyaml>=6.0.1
psutil>=5.9.0
